l2dataingest/scripts/requirements.txt
cx-Oracle>=8.2.1
pika>=1.2.1
python-dateutil>=2.8.1
PyYAML>=6.0
kafka-python>=2.0.2
flask>=2.1.3